+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - v3.2 -> v3.3 업그레이드
 - Activity.class new version 패널 액티비티 뷰 무한 생성 및 재활용, 에니메이션 등록 기능
 - 전쳬 예제 업데이트
+- Activity.clas url object type 으로 업데이트
 
 ### - 2024-05-31
 - ElementDiv,ElementUL,ElementLi style 메소드 기능 추가